+++ Modules/booking/Controller/BookingschedulingController.php
@@ -9,6 +9,8 @@ require_once 'Modules/resources/Model/ResourcesTranslator.php';
 require_once 'Modules/booking/Model/BkScheduling.php';
 require_once 'Modules/booking/Model/BkColorCode.php';
 require_once 'Modules/resources/Model/ReArea.php';
+require_once 'Modules/resources/Model/ResourceInfo.php';
+
 require_once 'Modules/core/Model/CoreStatus.php';
 require_once 'Modules/booking/Controller/BookingsettingsController.php';
 
@@ -32,6 +34,56 @@ class BookingschedulingController extends BookingsettingsController {
         $modelArea = new ReArea();
         $areas = $modelArea->getForSpace($id_space);
 
+        $modelResource = new ResourceInfo();
+        $resources = $modelResource->getForSpace($id_space);
+
+        $bk = new BkScheduling();
+        $bkcalendars = $bk->getAll($id_space);
+        $bks = new BkResourceSchedule();
+        $defScheduling = $bks->getDefault($id_space);
+        if ($defScheduling == null) {
+            throw new PfmException("No default scheduling found", 500);
+        }
+        $def = $defScheduling['id_bkschedule'];
+        $bklist = [$def => "default"];
+        $resbklist = [0 => "parent area"];
+        foreach ($bkcalendars as $cal) {
+            $bklist[$cal['id']] = $cal['name'];
+            $resbklist[$cal['id']] = $cal['name'];
+        }
+
+        
+        $calendars = $bks->all($id_space);
+        $resourcesCalendars = [];
+        $reareasCalendars = [];
+        foreach ($calendars as $cal) {
+            if($cal['id_resource'] == 0 && $cal['id_rearea'] == 0) {
+                continue;
+            }
+            if($cal['id_resource'] > 0){
+                $resourcesCalendars[$cal['id_resource']] = $cal['id_bkschedule'];
+            }
+            if($cal['id_rearea'] > 0){
+                $reareasCalendars[$cal['id_rearea']] = $cal['id_bkschedule'];
+            }
+        }
+
+        $mareas = [];
+
+        foreach ($areas as $i => $r) {
+            $areas[$i]['calendar'] = $reareasCalendars[$r['id']] ?? $def;
+            $mareas[$areas[$i]['id']] = $areas[$i];
+        }
+
+        foreach ($resources as $i => $r) {
+            if(isset($resourcesCalendars[$r['id']])) {
+                $resources[$i]['calendar'] = $resourcesCalendars[$r['id']];
+            } else {
+                $resources[$i]['calendar'] = 0;
+            }
+            $mareas[$resources[$i]['id_area']]['resources'][] = $resources[$i];
+        }
+
         if (empty($areas)) {
             $_SESSION['flash'] = ResourcesTranslator::Area_Needed($lang);
             $_SESSION['flashClass'] = "warning";
@@ -45,16 +97,17 @@ class BookingschedulingController extends BookingsettingsController {
         
         $tableHtml = $table->view($areas, $headers);
         
-        $this->render(array("id_space" => $id_space, "lang" => $lang, "tableHtml" => $tableHtml));
+        $this->render(array("id_space" => $id_space, "lang" => $lang, "tableHtml" => $tableHtml, "areas" => $mareas, "calendars" => $bklist, 'rescalendars' => $resbklist));
     }
     
     public function editAction($id_space, $id) {
         // That to avoid some confusions between id_rearea and bkScheduling['id']
-        $id_rearea = $id;
+        // $id_rearea = $id;
         
         $this->checkAuthorizationMenuSpace("bookingsettings", $id_space, $_SESSION["id_user"]);
         $lang = $this->getLanguage();
         
+        /*
         $modelArea = new ReArea();
         $area = $modelArea->get($id_space, $id_rearea);
         if (!$area) {
@@ -62,16 +115,19 @@ class BookingschedulingController extends BookingsettingsController {
         }
         
         $name = $area['name'];
+        */
         
         $modelScheduling = new BkScheduling();
-        $bkScheduling = $modelScheduling->getByReArea($id_space, $id_rearea);
-        if ($bkScheduling['id_rearea'] == 0) {
-            $bkScheduling['id_rearea'] = $id_rearea;
+        $bkScheduling = $modelScheduling->get($id_space, $id);
+        if ($bkScheduling['id_space'] == 0) {
             $bkScheduling["id_space"] = $id_space;
         }
+        $name = $bkScheduling['name'];
 
         $form = new Form($this->request, "bookingschedulingedit");
         $form->setTitle(BookingTranslator::Edit_scheduling($lang) . ": " . $name, 3);
+        $form->addHidden("id", $bkScheduling['id'] ?? 0);
+        $form->addText("name", CoreTranslator::Name($lang), true, $bkScheduling['name'] ?? '');
         $form->addChoicesList(BookingTranslator::Availables_days($lang),
                 BookingTranslator::DaysList($lang), 
                 array("is_monday", "is_tuesday", "is_wednesday", "is_thursday", "is_friday", "is_saturday", "is_sunday"),
@@ -112,7 +168,7 @@ class BookingschedulingController extends BookingsettingsController {
         $form->addSelectMandatory("default_color_id", BookingTranslator::Default_color($lang), $cc, $ccid, $bkScheduling["default_color_id"]);
         
         $todo = $this->request->getParameterNoException('redirect');
-        $validationUrl = "bookingschedulingedit/".$id_space."/".$id_rearea;
+        $validationUrl = "bookingschedulingedit/".$id_space."/".$id;
         if ($todo) {
             $validationUrl .= "?redirect=todo";
         }
@@ -132,7 +188,9 @@ class BookingschedulingController extends BookingsettingsController {
 
 
         if ($form->check()) {
-            $id_bkScheduling = $modelScheduling->edit($id_space, $bkScheduling['id_rearea'],
+            $id_bkScheduling = $modelScheduling->edit($id_space,
+                    $this->request->getParameterNoException("id"), 
+                    $this->request->getParameterNoException("name"), 
                     $this->request->getParameterNoException("is_monday"), 
                     $this->request->getParameterNoException("is_tuesday"), 
                     $this->request->getParameterNoException("is_wednesday"), 
@@ -156,7 +214,7 @@ class BookingschedulingController extends BookingsettingsController {
             if ($todo) {
                 return $this->redirect("spaceadminedit/" . $id_space, ["showTodo" => true]);
             } else {
-                return $this->redirect("bookingschedulingedit/".$id_space."/".$id_rearea, [], ['bkScheduling' => ['id' => $id_bkScheduling]]);
+                return $this->redirect("bookingschedulingedit/".$id_space."/".$id_bkScheduling, [], ['bkScheduling' => ['id' => $id_bkScheduling]]);
             }
             
         }
@@ -168,5 +226,29 @@ class BookingschedulingController extends BookingsettingsController {
             "data" => ["bkScheduling" => $bkScheduling]
         ));
         
+    }
+
+    public function assignAction($kind, $id_space, $id, $id_calendar) {
+        $bks = new BkResourceSchedule();
+        $link = 0;
+        if ($kind == 0) {
+            // area
+            $link = $bks->linkArea($id_space, $id, $id_calendar);
+        } else if ($kind == 1) {
+            // resource
+            if ($id_calendar == 0) {
+                $bks->unlinkResource($id_space, $id);
+            } else {
+                $link = $bks->linkResource($id_space, $id, $id_calendar);
+            }
+        }
+
+        return $this->render(['data' => ['id' => $link]]);
+    }
+
+    public function deleteAction($id_space, $id) {
+        $bk= new BkScheduling();
+        $bk->delete($id_space, $id);
+        $this->redirect('/bookingscheduling/'.$id_space);
     }
 }