+++ Modules/booking/Model/BkScheduling.php
@@ -2,18 +2,126 @@
 
 require_once 'Framework/Model.php';
 
+class BkResourceSchedule extends Model {
+
+    public function __construct() {
+        $this->tableName = "bk_resource_schedule";
+    }
+
+    public function createTable() {
+        $sql = "CREATE TABLE IF NOT EXISTS `bk_resource_schedule` (
+		    `id` int NOT NULL AUTO_INCREMENT,
+            `id_space` int NOT NULL,
+            `id_bkschedule` int NOT NULL,
+            `id_rearea` int NOT NULL DEFAULT 0,
+            `id_resource` int NOT NULL DEFAULT 0,
+		    PRIMARY KEY (`id`)
+		);";
+        $this->runRequest($sql);
+        $this->baseSchema();
+    }
+
+    public function unlinkResource($id_space, $id_resource){
+        $sql = 'DELETE FROM bk_resource_schedule WHERE id_space=? AND id_resource=?';
+        $this->runRequest($sql, [$id_space, $id_resource]);
+    }
+
+    public function unlinkArea($id_space, $id_rearea){
+        $sql = 'DELETE FROM bk_resource_schedule WHERE id_space=? AND id_rearea=?';
+        $this->runRequest($sql, [$id_space, $id_rearea]);
+    }
+
+    public function unlinkAll($id_space, $id_bkschedule){
+        $sql = 'DELETE FROM bk_resource_schedule WHERE id_space=? AND id_bkschedule=?';
+        $this->runRequest($sql, [$id_space, $id_bkschedule]);
+    }
+
+    public function linkResource($id_space, $id_resource, $id_bkschedule){
+        $this->unlinkResource($id_space, $id_resource);
+        $sql = 'INSERT INTO bk_resource_schedule (id_space, id_bkschedule, id_resource) VALUES (?, ?, ?)';
+        $this->runRequest($sql, [$id_space, $id_bkschedule, $id_resource]);
+        return $this->getDatabase()->lastInsertId();
+    }
+
+    public function all($id_space) {
+        $sql = 'SELECT bk_resource_schedule.*, bk_schedulings.name FROM bk_resource_schedule INNER JOIN bk_schedulings ON bk_schedulings.id=bk_resource_schedule.id_bkschedule WHERE bk_resource_schedule.id_space=?';
+        return $this->runRequest($sql, [$id_space])->fetchAll();
+    }
+
+    public function linkedResource($id_space, $id_bkschedule) {
+        $sql = 'SELECT * from bk_resource_schedule WHERE id_bkschedule=? AND id_space=? AND id_resource>0';
+        return $this->runRequest($sql, [$id_bkschedule, $id_space])->fetchAll();
+    }
+
+    public function linkedArea($id_space, $id_bkschedule) {
+        $sql = 'SELECT * from bk_resource_schedule WHERE id_bkschedule=? AND id_space=? AND id_rearea>0';
+        return $this->runRequest($sql, [$id_bkschedule, $id_space])->fetchAll();
+    }
+
+    public function linkArea($id_space, $id_rearea, $id_bkschedule){
+        $this->unlinkArea($id_space, $id_rearea);
+        $sql = 'INSERT INTO bk_resource_schedule (id_space, id_bkschedule, id_rearea) VALUES (?, ?, ?)';
+        $this->runRequest($sql, [$id_space, $id_bkschedule, $id_rearea]);
+        return $this->getDatabase()->lastInsertId();
+    }
+
+    public function setDefault($id_space, $id_bkschedule){
+        $sql = 'INSERT INTO bk_resource_schedule (id_space, id_bkschedule) VALUES (?, ?)';
+        $this->runRequest($sql, [$id_space, $id_bkschedule]);
+    }
+
+    public function getDefault($id_space){
+        $sql = 'SELECT * from bk_resource_schedule WHERE id_space=? AND id_resource=0 AND id_rearea=0';
+        $res = $this->runRequest($sql, [$id_space]);
+        if($res->rowCount() == 1) {
+            return $res->fetch();
+        }
+        return null;
+    }
+
+    /**
+     * Get bkscheduling id for resource
+     * If none, get the one defined for resource area
+     * If none, get default one
+     * Else return null
+     */
+    public function getResourceScheduling($id_space, $id_resource) {
+        $sql = 'SELECT * from bk_resource_schedule WHERE id_space=? AND id_resource=?';
+        $res = $this->runRequest($sql, [$id_space, $id_resource]);
+        if($res->rowCount() == 1) {
+            return $res->fetch();
+        }
+
+        $sql = 'SELECT id_area FROM re_info WHERE deleted=0 AND id_space=? AND id=?';
+        $res = $this->runRequest($sql, [$id_space, $id_resource]);
+        if($res->rowCount() == 1) {
+            $area = $res->fetch();
+            $id_rearea = $area['id_area'];
+            $sql = 'SELECT * from bk_resource_schedule WHERE id_space=? AND id_rearea=?';
+            $res = $this->runRequest($sql, [$id_space, $id_rearea]);
+            if($res->rowCount() == 1) {
+                return $res->fetch();
+            }
+        }
+
+        $sql = 'SELECT * from bk_resource_schedule WHERE id_space=? AND id_rearea=0 AND id_resource=0';
+        $res = $this->runRequest($sql, [$id_space]);
+        if($res->rowCount() == 1) {
+            return $res->fetch();
+        }
+        return null;
+        
+    }
+}
+
+
 /**
- * Class defining the SyColorCode model
+ * Class defining the booking scheduling model
  *
  * @author Sylvain Prigent
  */
 class BkScheduling extends Model {
 
-    /**
-     * Create the SyColorCode table
-     * 
-     * @return PDOStatement
-     */
     public function __construct() {
 
         $this->tableName = "bk_schedulings";
@@ -31,15 +139,25 @@ class BkScheduling extends Model {
         $this->setColumnsInfo("booking_time_scale", "int(5)", 1);
         $this->setColumnsInfo("resa_time_setting", "int(1)", 1);
         $this->setColumnsInfo("default_color_id", "int(11)", 1);
-        $this->setColumnsInfo("id_rearea", "int(11)", 0);
+        // $this->setColumnsInfo("id_rearea", "int(11)", 0);
         $this->setColumnsInfo("force_packages", "tinyint", 0);
         $this->setColumnsInfo('shared', 'tinyint', 0);
+        $this->setColumnsInfo("name", "varchar(100)", "");
         $this->primaryKey = "id";
     }
 
-    public function getDefault() {
+    public function getDefault($id_space=0) {
+        if($id_space > 0) {
+            $bks = new BkResourceSchedule();
+            $def = $bks->getDefault($id_space);
+            if($def) {
+                return $def;
+            }
+        }
         return array(
             "id" => 0,
+            "id_space" => $id_space,
+            "name" => "",
             "is_monday" => 1,
             "is_tuesday" => 1,
             "is_wednesday" => 1,
@@ -53,7 +171,7 @@ class BkScheduling extends Model {
             "booking_time_scale" => 2,
             "resa_time_setting" => 2,
             "default_color_id" => 1,
-            "id_rearea" => 0,
+            // "id_rearea" => 0,
             "force_packages" => 0,
             "shared" => 0
         );
@@ -74,8 +192,14 @@ class BkScheduling extends Model {
         if($minutes == "") {
             $minutes = 0;
         }
-        $sql = "SELECT size_bloc_resa FROM bk_schedulings WHERE id_rearea=(SELECT id_area FROM re_info WHERE id=? AND deleted=0 AND id_space=? )";
-        $req = $this->runRequest($sql, array($id_resource, $id_space));
+        $bks = new BkResourceSchedule();
+        $sched = $bks->getResourceScheduling($id_space, $id_resource);
+        if ($sched == null) {
+            return $this->getDefault()['size_bloc_resa'];
+        }
+        $sql = "SELECT size_bloc_resa FROM bk_schedulings WHERE id=? AND id_space=?";
+        // $sql = "SELECT size_bloc_resa FROM bk_schedulings WHERE id_rearea=(SELECT id_area FROM re_info WHERE id=? AND deleted=0 AND id_space=? )";
+        $req = $this->runRequest($sql, array($sched['id'], $id_space));
         if ($req->rowCount() > 0){
             $d = $req->fetch();
             $step = $d[0];
@@ -95,9 +219,9 @@ class BkScheduling extends Model {
     }
     
     /**
-     * get SyColorCodes informations
+     * get all calendars
      * 
-     * @param string $sortentry Entry that is used to sort the SyColorCodes
+     * @param string $sortentry Entry that is used to sort
      * @return multitype: array
      */
     public function getAll($id_space, $sortentry = 'id') {
@@ -123,52 +247,62 @@ class BkScheduling extends Model {
         return $user->fetch();
     }
 
+    public function getByResource($id_space, $id_resource) {
+        $bks = new BkResourceSchedule();
+        $sched = $bks->getResourceScheduling($id_space, $id_resource);
+
+        if ($sched == null) {
+            return $this->getDefault();
+        }
+        $sql = "SELECT * FROM bk_schedulings WHERE id=? AND deleted=0 AND id_space=?";
+        $res = $this->runRequest($sql, array($sched['id_bkschedule'], $id_space));
+        $scheduling = $res->fetch();
+
+        if (!$scheduling) {
+            $scheduling = $this->getDefault($id_space);
+        }
+        return $scheduling;
+    }
+
+    /*
     public function getByReArea($id_space, $id_rearea) {
         $sql = "SELECT * FROM bk_schedulings WHERE id_rearea=? AND deleted=0 AND id_space=?";
-        $user = $this->runRequest($sql, array($id_rearea, $id_space));
-        $scheduling = $user->fetch();
+        $res = $this->runRequest($sql, array($id_rearea, $id_space));
+        $scheduling = $res->fetch();
         if (!$scheduling) {
             $scheduling = $this->getDefault();
         }
         return $scheduling;
     }
+    */
 
     /**
-     * add a SyColorCode to the table
+     * add a schedule to the table
      *
-     * @param string $name name of the SyColorCode
-     * @param string $address address of the SyColorCode
      */
-    public function add($id_space, $id_rearea, $is_monday, $is_tuesday, $is_wednesday, $is_thursday, $is_friday, $is_saturday, $is_sunday, $day_begin, $day_end, $size_bloc_resa, $booking_time_scale, $resa_time_setting, $default_color_id, $force_packages=0, $shared=0) {
+    public function add($id_space, $name, $is_monday, $is_tuesday, $is_wednesday, $is_thursday, $is_friday, $is_saturday, $is_sunday, $day_begin, $day_end, $size_bloc_resa, $booking_time_scale, $resa_time_setting, $default_color_id, $force_packages=0, $shared=0) {
 
-        $sql = "insert into bk_schedulings(is_monday, is_tuesday, "
+        $sql = "insert into bk_schedulings(name, is_monday, is_tuesday, "
                 . " is_wednesday, is_thursday, is_friday, is_saturday, is_sunday, day_begin,"
                 . " day_end, size_bloc_resa, booking_time_scale,"
-                . " resa_time_setting, default_color_id, id_space, id_rearea, force_packages, shared)"
+                . " resa_time_setting, default_color_id, id_space, force_packages, shared)"
                 . " values(?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)";
-        $this->runRequest($sql, array($is_monday, $is_tuesday,
+        $this->runRequest($sql, array($name, $is_monday, $is_tuesday,
             $is_wednesday, $is_thursday, $is_friday, $is_saturday, $is_sunday, $day_begin,
             $day_end, $size_bloc_resa, $booking_time_scale,
-            $resa_time_setting, $default_color_id, $id_space, $id_rearea, $force_packages, $shared));
+            $resa_time_setting, $default_color_id, $id_space, $force_packages, $shared));
         return $this->getDatabase()->lastInsertId();
     }
 
-    /**
-     * update the information of a SyColorCode
-     *
-     * @param int $id Id of the SyColorCode to update
-     * @param string $name New name of the SyColorCode
-     * @param string $color New Address of the SyColorCode
-     */
-    public function update2($id_space, $id_rearea, $is_monday, $is_tuesday, $is_wednesday, $is_thursday, $is_friday, $is_saturday, $is_sunday, $day_begin, $day_end, $size_bloc_resa, $booking_time_scale, $resa_time_setting, $default_color_id, $force_packages=0, $shared=0) {
+    public function bkupdate($id_space, $id, $name, $is_monday, $is_tuesday, $is_wednesday, $is_thursday, $is_friday, $is_saturday, $is_sunday, $day_begin, $day_end, $size_bloc_resa, $booking_time_scale, $resa_time_setting, $default_color_id, $force_packages=0, $shared=0) {
 
-        $sql = "UPDATE bk_schedulings SET is_monday=?, is_tuesday=?, is_wednesday=?, is_thursday=?, is_friday=?, "
+        $sql = "UPDATE bk_schedulings SET name=?, is_monday=?, is_tuesday=?, is_wednesday=?, is_thursday=?, is_friday=?, "
                 . "is_saturday=?, is_sunday=?, day_begin=?, day_end=?, size_bloc_resa=?, booking_time_scale=?, "
                 . "resa_time_setting=?, default_color_id=?, force_packages=?, shared=? "
-                . "WHERE id_rearea=? AND deleted=0 AND id_space=?";
-        $this->runRequest($sql, array($is_monday, $is_tuesday, $is_wednesday, $is_thursday, $is_friday,
+                . "WHERE id=? AND deleted=0 AND id_space=?";
+        $this->runRequest($sql, array($name, $is_monday, $is_tuesday, $is_wednesday, $is_thursday, $is_friday,
             $is_saturday, $is_sunday, $day_begin, $day_end, $size_bloc_resa,
-            $booking_time_scale, $resa_time_setting, $default_color_id, $force_packages, $shared, $id_rearea, $id_space));
+            $booking_time_scale, $resa_time_setting, $default_color_id, $force_packages, $shared, $id, $id_space));
     }
 
     protected function onToBool($on){
@@ -183,7 +317,8 @@ class BkScheduling extends Model {
     
     public function edit(
         $id_space,
-        $id_rearea,
+        $id,
+        $name,
         $is_monday,
         $is_tuesday,
         $is_wednesday,
@@ -200,11 +335,11 @@ class BkScheduling extends Model {
         $force_packages,
         $shared
     ) {
-        $id = 0;
-        if ($id = $this->existsByReArea($id_space, $id_rearea)) {
-            $this->update2(
+        if ($id > 0) {
+            $this->bkupdate(
                 $id_space,
-                $id_rearea,
+                $id,
+                $name,
                 $this->onToBool($is_monday),
                 $this->onToBool($is_tuesday),
                 $this->onToBool($is_wednesday),
@@ -224,7 +359,7 @@ class BkScheduling extends Model {
         } else {
             $id = $this->add(
                 $id_space,
-                $id_rearea,
+                $name,
                 $this->onToBool($is_monday),
                 $this->onToBool($is_tuesday),
                 $this->onToBool($is_wednesday),
@@ -256,36 +391,66 @@ class BkScheduling extends Model {
         return ($req->rowCount() == 1);
     }
 
+    public function nameExists($id_space, $name){
+        $sql = "SELECT id FROM bk_schedulings WHERE id_space=? AND name=?";
+        $res = $this->runRequest($sql, [$id_space, $name]);
+        if($res->rowCount()>0) {
+            return true;
+        }
+        return false;
+    }
+
     public function getForSpace($id_space) {
         $sql = "SELECT * from bk_schedulings WHERE id_space=? AND deleted=0";
         return $this->runRequest($sql, array($id_space))->fetchAll();
     }
 
     /**
-     * Check if a bkScheduling exists from area
-     * @param string $id_space
-     * @param string $id_rearea
-     * @return boolean
-     */
-    public function existsByReArea($id_space, $id_rearea): ?array {
-        $sql = "SELECT * from bk_schedulings WHERE id_rearea=? AND deleted=0 AND id_space=?";
-        $req = $this->runRequest($sql, array($id_rearea, $id_space));
-        return ($req->rowCount() == 1) ? $req->fetch() : null;
-    }
-
-    /**
-     * Remove a color code
+     * Remove a schedule
      * @param unknown $id
      */
     public function delete($id_space, $id) {
+        $def = $this->getDefault($id_space);
+        if($def['name'] == 'default'){
+            throw new PfmParamException('Cannot delete default scheduling');
+        }
         $sql = "UPDATE bk_schedulings SET deleted=1,deleted_at=NOW() WHERE id=? AND id_space=?";
-        //$sql = "DELETE FROM bk_schedulings WHERE id = ? AND deleted=0 AND id_space=?";
         $this->runRequest($sql, array($id, $id_space));
+        $bks = new BkResourceSchedule();
+        $bks->unlinkAll($id_space, $id);
     }
 
-    public function setReArea($id_space, $id, $id_rearea) {
-        $sql = "UPDATE bk_schedulings SET id_rearea=? WHERE id=? AND deleted=0 AND id_space=?";
-        $this->runRequest($sql, array($id_rearea, $id, $id_space));
+    public function createDefault($id_space){
+        $default = $this->getDefault($id_space);
+        if ($default['id']) {
+            Configuration::getLogger()->warning('[scheduling][create default] a default already exists', ['id' => $default['id']]);
+            return $default['id'];
+        }
+        Configuration::getLogger()->warning('[scheduling][create default]', ['id_space' => $id_space, 'default' => $default]);
+        $id = $this->add(
+            $id_space,
+            'default',
+            $default['is_monday'],
+            $default['is_tuesday'],
+            $default['is_wednesday'],
+            $default['is_thursday'],
+            $default['is_friday'],
+            $default['is_saturday'],
+            $default['is_sunday'],
+            $default['day_begin'],
+            $default['day_end'],
+            $default['size_bloc_resa'],
+            $default['booking_time_scale'],
+            $default['resa_time_setting'],
+            $default['default_color_id'],
+            $default['force_packages'],
+            $default['shared']
+        );
+        if($id) {
+            $bks = new BkResourceSchedule();
+            $bks->setDefault($id_space, $id);
+        }
+        return $id;
     }
 
 }