+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features / enhancements
 
 * [cli] add maintenance option to command-line admin tool
+* [booking] support multiple calendars
 * [core] add support for custom plans
 * [rating] add new feature satisfaction campaigns #616
 * [booking] make confirmation email optional at booking